+++ src/mruby_integration/models/image.cpp
@@ -73,6 +73,22 @@ mrb_Image_unload(mrb_state* mrb, mrb_value self) -> mrb_value
 }
 
 auto
+mrb_Image_export(mrb_state* mrb, mrb_value self) -> mrb_value
+{
+  Image* image;
+
+  Data_Get_Struct(mrb, self, &Image_type, image);
+  mrb_assert(image != nullptr);
+
+  char* path;
+  mrb_get_args(mrb, "z", &path);
+
+  ExportImage(*image, path);
+
+  return mrb_nil_value();
+}
+
+auto
 mrb_Image_get_data(mrb_state* mrb, mrb_value self) -> mrb_value
 {
   Image* image;
@@ -112,6 +128,8 @@ append_models_Image(mrb_state* mrb)
     mrb, Image_class, "initialize", mrb_Image_initialize, MRB_ARGS_REQ(5));
   mrb_define_method(
     mrb, Image_class, "unload", mrb_Image_unload, MRB_ARGS_NONE());
+  mrb_define_method(
+    mrb, Image_class, "export", mrb_Image_export, MRB_ARGS_REQ(1));
   mrb_define_method(
     mrb, Image_class, "data", mrb_Image_get_data, MRB_ARGS_NONE());